backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
google-play-scraper==1.2.4
aiohttp==3.8.6
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
